Added Base Django (with Mezzanine) Framework [BUGGY]
To start server, move into */website/ then run from terminal "python
manage.py runserver" (This should work without having mezzanine
installed but I haven't tested it. If it doesn't install mezzanine via
"pip install mezzanine")

The default login details for now are user: admin, pass: pass.

Note that the homepage will throw an error and not render for the time
being, as I'm having issues getting the mezzanine blog template to
render when targeted as a local path copy. In order to get it to render
with the "default" home page that comes with mezzanine, change line 36
in urls.py (located in */website/) to "url("^$", direct_to_template,
{"template": "index.html"}, name="home"),"
To see what it /should/ render as with the blog components appearing on
the front page (granted this will all be stock mezzanine styling and
not any of the few layout edits I've made so far) comment out line 36
and uncomment line 61 of urls.py.

I think that's about everything, but if anything's not working or you
have queries on how to do something just let me know and I'll hopefully
have an answer (I've spent a few solid days trying to wrap my head
around everything so fingers crossed).