CSS Edits, README.md updated
Updated the readme. Have a look at it if you have the chance to see
what the plans are from now until ship date (aiming for Wednesday 19th
at this stage)

Began to fix up the CSS. Gotten the top navbar section almost done
(obvious issues are font sizes, as well as hyperlink dramas as detailed
in readme) as well as the footer, and the calendar widget. It's
starting to resemble the final site form which is good. Not too much
left now.